web/fastapi/requirements.txt
# FastAPI依赖列表
# 使用方法: pip install -r requirements.txt

# ============================================
# FastAPI核心框架
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6

# ============================================
# 认证和安全
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.0

# ============================================
# 数据库
# ============================================
pymysql==1.1.0
mysql-connector-python==8.3.0
cryptography==44.0.0
sqlalchemy==2.0.36

# ============================================
# 配置管理
# ============================================
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0

# ============================================
# 异步支持
# ============================================
aiofiles==23.2.1
httpx==0.28.0
aiohttp==3.11.0

# ============================================
# AI和机器学习
# ============================================
# OpenAI API
openai==2.21.0

# Anthropic API
anthropic==0.83.0

# 机器学习基础库
numpy>=1.24.0
pandas>=2.2.0
scikit-learn>=1.3.0  
joblib>=1.3.0        
xgboost>=2.0.0       
scipy>=1.10.0        



# 定时任务
apscheduler>=3.10.0

# ============================================
# 数据处理和可视化
# ============================================
requests==2.32.3
beautifulsoup4==4.14.3
lxml==5.1.0
tqdm==4.67.3

# ============================================
# 文件分析工具
# ============================================
pefile>=2023.2.7
h5py>=3.10.0

# ============================================
# 其他工具
# ============================================
python-dateutil==2.8.2
pytz==2024.1
